requirements.txt
ultralytics>=8.3.0
albumentations>=1.4.0
opencv-python>=4.9.0
# NumPy 2.x can break older PyTorch/OpenCV wheels; stay on 1.x for stability.
numpy>=1.26.0,<2.0
matplotlib>=3.8.0
Pillow>=10.2.0
tqdm>=4.66.0
pyyaml>=6.0.1
scikit-learn>=1.4.0
torch>=2.2.0
torchvision>=0.17.0
grad-cam>=1.5.0
tensorboard>=2.16.0
seaborn>=0.13.0
pandas>=2.2.0
rich>=13.7.0
